+++ test/specs/code/modules.js
@@ -1019,7 +1019,6 @@ a span {
                     pattern: "[local]-name-[name]-folder-[folder]-ext-[ext]-path-[path]-hash-[hash:base64:5]",
                 },
             }).then((result) => {
-
                 expect(result.code)
                     .equals(`.button-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-YnV0d {
  background-color: #007bff;
@@ -1066,6 +1065,122 @@ a span {
  background: red
 }`);
             });
+        });
+
+        it("module pattern #27", function () {
+            const file = new URL(dirname(import.meta.url) + "/../../css-modules/button.css");
+
+            const result = transformSync({
+                src: file.pathname,
+                input: `/* Button.module.css file */
+
+.button {
+    background-color: #007bff;
+    color: #ffffff;
+    padding: 10px 20px;
+    border: none;
+    cursor: pointer;
+    border-radius: 4px;
+}
+
+.button:hover {
+    background-color: #0056b3;
+}
+
+@property --progress {
+    syntax: "<percentage>";
+    inherits: false;
+    initial-value: 25%;
+}
+
+.bar {
+    display: inline-block;
+    --progress: 25%;
+    width: 100%;
+    height: 5px;
+    background: linear-gradient(
+            to right,
+            #00d230 var(--progress),
+            black var(--progress)
+    );
+    animation: progressAnimation 2.5s ease infinite;
+}
+
+@keyframes progressAnimation {
+    to {
+        --progress: 100%;
+    }
+}
+
+.body {
+    background: #6e28d9;
+    padding: 0 24px;
+    color: white; /* Change my color to yellow */
+    margin: 0;
+    height: 100vh;
+    justify-content: center;
+    align-items: center;
+
+}
+
+.animation {
+    display: block;
+    width: var(--progress);
+    animation: progressAnimation infinite alternate 3s;
+    background: red;
+}
+`,
+                beautify: true,
+                module: {
+                    pattern: "[local]-name-[name]-folder-[folder]-ext-[ext]-path-[path]-hash-[hash:base64:5]",
+                },
+            });
+
+            expect(result.code)
+                .equals(`.button-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-YnV0d {
+ background-color: #007bff;
+ color: #fff;
+ padding: 10px 20px;
+ border: 0;
+ cursor: pointer;
+ border-radius: 4px
+}
+.button-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-YnV0d:hover {
+ background-color: #0056b3
+}
+@property --progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ {
+ syntax: "<percentage>";
+ inherits: false;
+ initial-value: 25%
+}
+.bar-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-YmFyO {
+ display: inline-block;
+ --progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ: 25%;
+ width: 100%;
+ height: 5px;
+ background: linear-gradient(90deg,#00d230 var(--progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ),#000 var(--progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ));
+ animation: progressAnimation-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ 2.5s infinite
+}
+@keyframes progressAnimation-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ {
+ to {
+  --progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ: 100%
+ }
+}
+.body-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-Ym9ke {
+ background: #6e28d9;
+ padding: 0 24px;
+ color: #fff;
+ margin: 0;
+ height: 100vh;
+ justify-content: center;
+ align-items: center
+}
+.animation-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-YW5pb {
+ display: block;
+ width: var(--progress-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ);
+ animation: progressAnimation-name-button-folder-css-modules-ext-css-path-test_css-modules_button_css-hash-cHJvZ infinite alternate 3s;
+ background: red
+}`);
         });
     });
 }